Move date_grid to ui forlder (#13986)
