+++ js/renovation/ui/data_grid/__tests__/data_grid.test.tsx
@@ -1,14 +1,14 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { h, createRef } from 'preact';
 import { mount, shallow } from 'enzyme';
-import DxDataGrid from '../../../ui/data_grid';
+import DxDataGrid from '../../../../ui/data_grid';
 import { viewFunction as DataGridView, DataGrid } from '../data_grid';
 import { DataGridProps } from '../props';
 
 const mockDispose = jest.fn();
 const mockOption = jest.fn();
 
-jest.mock('../../../ui/data_grid', () => {
+jest.mock('../../../../ui/data_grid', () => {
   const MockDxDataGrid = jest.fn().mockImplementation(() => ({
     dispose: mockDispose,
     option: mockOption,
